fixed unwanted state persistance loading different data between same component, added new reducer store for useStories

+++ client/src/components/Story.js
@@ -1,16 +1,20 @@
 import React, { useEffect } from 'react';
 import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
-import { fetchStory, fetchAuthor, deleteStory } from '../actions';
+import { fetchStory, fetchAuthor, deleteStory, cleanUp } from '../actions';
 
 import StoryComments from './StoryComments';
 
-const Story = ({ match, story, fetchStory, fetchAuthor, author, deleteStory, auth, history }) => {
+const Story = ({ match, story, fetchStory, fetchAuthor, author, deleteStory, auth, history, cleanUp }) => {
 
 
 
   useEffect(() => {
     fetchStory(match.params.storyId);
+
+    return function cleanup() {
+
+    }
   }, [fetchStory, match.params.storyId]);
 
   useEffect(() => {
@@ -18,6 +22,13 @@ const Story = ({ match, story, fetchStory, fetchAuthor, author, deleteStory, aut
       fetchAuthor(story._user);
     }
   }, [story, fetchAuthor])
+
+  useEffect(() => {
+
+    return function cleanup() {
+      cleanUp()
+    }
+  }, [])
 
 
   const authorNameCheck = (author) => {
@@ -34,17 +45,43 @@ const Story = ({ match, story, fetchStory, fetchAuthor, author, deleteStory, aut
 
 
 
+  const renderEditOrDonate = () => {
+    if (story && auth) {
+      if (story._user === auth._id) {
+        return (
+          <div className="story-btns-container">
+            <Link className="btn" to={`/edit-story/${story._id}`}>Edit</Link>
+            <button
+              className="btn"
+              onClick={() => onDelete(story._id)}
+            >delete</button>
+          </div>
+        )
+      }
+      else {
+        return (
+          <div className="story-btns-container">
+            <p>Like the story?</p>
+            <Link
+              to={`/payment/${author._id}`}
+              className="btn donate-btn"
+            >Donate
+            </Link>
+          </div>
+        )
+      }
+    }
+  }
+
+
+
+
   return (
     <div>
       {author && story ?
         <div className="story-page-container">
           <h3>{story.title}</h3>
-          <Link className="btn" to={`/edit-story/${story._id}`}>Edit</Link>
-          <button
-            className="btn"
-            onClick={() => onDelete(story._id)}
-          >delete</button>
-
+          {renderEditOrDonate()}
           <p>{story.creator}</p>
           <p className="story-page-desc">{story.description}</p>
           <div className="story-page-grid">
@@ -58,10 +95,7 @@ const Story = ({ match, story, fetchStory, fetchAuthor, author, deleteStory, aut
                 <h4>Written By:
                    <Link to={`/author/${story._user}`}><span>{authorNameCheck(author)}</span></Link>
                 </h4>
-                <Link
-                  to={`/payment/${author._id}`}
-                  className="btn donate-btn"
-                >Donate</Link>
+
               </div>
               <StoryComments />
             </div>
@@ -76,4 +110,4 @@ const Story = ({ match, story, fetchStory, fetchAuthor, author, deleteStory, aut
 const mapStateToProps = ({ story, author, auth }) => {
   return { story, author, auth }
 }
-export default connect(mapStateToProps, { fetchStory, fetchAuthor, deleteStory })(Story)
+export default connect(mapStateToProps, { fetchStory, fetchAuthor, deleteStory, cleanUp })(Story)
